2020-1/qp1/question-4/requirements.txt
astroid==2.4.2
autopep8==1.5.3
glfw==1.11.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multipledispatch==0.6.0
numpy==1.18.5
Pillow==7.1.2
pycodestyle==2.6.0
pygame==1.9.6
pylint==2.5.3
PyOpenGL==3.1.5
pyrr==0.10.3
six==1.15.0
toml==0.10.1
wrapt==1.12.1
